+++ test/interactive-swap-test.js
@@ -83,7 +83,7 @@ describe('Interactive name swap', function() {
   });
 
   it('should win name with Alice\'s wallet', async () => {
-    await alice.sendOpen(name, false);
+    await alice.sendOpen(name);
     await mineBlocks(network.names.treeInterval + 1);
 
     await alice.sendBid(name, 100000, 200000);